+++ hosts/default/variables.nix
@@ -52,6 +52,7 @@
   terminal = "kitty"; # Set Default System Terminal
 
   keyboardLayout = "us";
+  keyboardVariant = "";
   consoleKeyMap = "us";
 
   # For hybrid support (Intel/NVIDIA Prime or AMD/NVIDIA)