+++ hosts/meo/default.nix
@@ -4,6 +4,8 @@
     ./host-packages.nix
   ];
 
+  programs.kdeconnect.enable = true;
+
   # Vivaldi als Standard-Browser erzwingen
   xdg.mime.defaultApplications = {
     "text/html" = "vivaldi-stable.desktop";