+++ modules/core/xserver.nix
@@ -1,13 +1,18 @@
 { host, ... }:
 let
-  inherit (import ../../hosts/${host}/variables.nix) keyboardLayout;
+  vars = import ../../hosts/${host}/variables.nix;
+  keyboardLayout = vars.keyboardLayout or "us";
+  keyboardVariant = vars.keyboardVariant or "";
+  layoutIsVariant = builtins.elem keyboardLayout [ "dvorak" "colemak" "workman" "intl" "us-intl" ];
+  xkbLayout = if (keyboardVariant != "" || layoutIsVariant) then "us" else keyboardLayout;
+  xkbVariant = if (keyboardVariant != "") then keyboardVariant else if layoutIsVariant then keyboardLayout else "";
 in
 {
   services.xserver = {
     enable = false;
     xkb = {
-      layout = "${keyboardLayout}";
-      variant = "";
+      layout = xkbLayout;
+      variant = xkbVariant;
     };
   };
 }