+++ CHANGELOG.md
@@ -5,13 +5,25 @@
 
 ---
 
-# 🚀 **Current Release - ZaneyOS v2.5.2**
+# 🚀 **Current Release - ZaneyOS v2.5.3**
+
+#### 📅 **Updated: December 3rd, 2025**
+
+- Added: Support for `kb_variants` at install time.
+  - Keyboards, like `dvorak`, `coleman,` or `euro`
+  - Configures, console, SDDM and Hyprland
+- Fixed: TUI greetd service name changed to `services.greetd`
 
 #### 📅 **Updated: December 1st, 2025**
 
+- Added: System keybinds search tool from ddubsOS
+  - Shows bindings for:
+  - Hyprland, kitty, wezterm, ghostty, emacs and yazi
+  - SUPER + K to invoke
+  - SUPER + SHIFT + K for legacy keybind rofi menu
 - Fixed: Added power profiles daemon for noctalia
-- Fixed: `gpu-screen-recorder` not installed 
-   - Thanks to `@mugdad11` for finding them 
+- Fixed: `gpu-screen-recorder` not installed
+  - Thanks to `@mugdad11` for finding them
 - Fixed: Both rofi and noctalia clipboard
   - Both were active on `SUPER + V`
   - Now checks for waybar or noctalia shell
@@ -77,8 +89,7 @@
     - gh
     - lazygit
 
-    Contributors: 
-   `@mugdad11`
+    Contributors: `@mugdad11`
 
 #### 📅 **Updated: September 18th, 2025**